Implements the skip statement parsing and execution
This commit also refactors the grammar so that we can test parts of it,
or the whole thing if we need to. This is especially useful since
plain literals are not valid Oz programms (which is a single
statement), so we can't test literal parsing if we don't have a
separate parser for that grammar.

This adds the basic infrastructure for kernelizing abstract syntax
trees, building an abstract machine out of a kernelized abstract
syntax tree and stepping through the execution of an abstract machine.

This also adds a bit of a console demo on how the machine works for the
web application itself.