Implement lexical syntax (#12)
* Refactors literals to be lexical

Values and variables are now inside a category of lexical syntactic
elements. All literal values are now inside this category. This also
implements type hierarchies, and regroups parser files into directories.

* Configures eslint properly

* Refactors lexical nodes into value nodes

This makes every composite value be parsed as records, including atoms
and boolean literals, and every number descendant to be parsed as a
number.

I also revamped the specs so that we get useful information messages on
complex comparissons errors.